+++ cypress/integration/collections_spec.js
@@ -27,6 +27,7 @@ describe('Dashboard Collections Page', () => {
       cy.server();
       cy.route('POST', '/collections').as('postCollection');
       cy.route('GET', '/collections?limit=*').as('getCollections');
+      cy.route('GET', '/collections/active?limit=*').as('getActiveCollections');
       cy.route('GET', '/collections?name=*').as('getCollection');
       cy.route('GET', '/granules?limit=*').as('getGranules');
 
@@ -39,22 +40,52 @@ describe('Dashboard Collections Page', () => {
       cy.contains('nav li a', 'Collections').as('collections');
       cy.get('@collections').should('have.attr', 'href', '/collections');
       cy.get('@collections').click();
-      cy.wait('@getCollections');
+      cy.wait('@getActiveCollections');
 
       cy.url().should('include', 'collections');
       cy.contains('.heading--xlarge', 'Collections');
 
+      cy.get('.table .tbody .tr').should('have.length', 1);
+
+      cy.clearStartDateTime();
+      cy.wait('@getCollections');
+
       cy.get('.table .tbody .tr').should('have.length', 5);
+    });
+
+    it('should only display collections with active granules when time filter is applied', () => {
+      cy.contains('nav li a', 'Collections').as('collections');
+      cy.get('@collections').should('have.attr', 'href', '/collections');
+      cy.get('@collections').click();
+      cy.wait('@getActiveCollections');
+
+      cy.url().should('include', 'collections');
+      cy.contains('.heading--xlarge', 'Collections');
+
+      cy.get('.table .tbody .tr').as('listItems');
+
+      cy.get('@listItems').each(($row) => {
+        // verify granule column does not equal 0
+        cy.wrap($row).find('.td').eq(3).should('not.eq', '0');
+      });
+
+      cy.get('@listItems').find('.td a').eq(0).click();
+      cy.wait('@getGranules');
+
+      // verify there is a granule with a timestamp containing second or minute
+      // this would indicate it was updated within the default timeframe of 1 hour
+      cy.get('@listItems').should('have.length', 11).contains('.td', /second|minute/);
     });
 
     it('should display expected MMT Links for collections list', () => {
       cy.visit('/collections');
+      cy.clearStartDateTime();
       cy.wait('@getCollections');
       let i = 0;
 
       cy.get('.table .tbody .tr').should('have.length', 5);
 
-      while (i < cmrFixtureIdx) cy.wait(`@cmr${i++}`, {timeout: 25000});
+      while (i < cmrFixtureIdx) cy.wait(`@cmr${i++}`, { timeout: 25000 });
       cy.contains('.table .tbody .tr', 'MOD09GQ')
         .contains('.td a', 'MMT')
         .should('have.attr', 'href')
@@ -75,6 +106,7 @@ describe('Dashboard Collections Page', () => {
       // On the Collections page, click the Add Collection button
       cy.visit('/collections');
       cy.contains('.heading--large', 'Collection Overview');
+      cy.clearStartDateTime();
       cy.contains('a', 'Add Collection').click();
 
       // Fill the form with the test collection JSON and submit it
@@ -120,13 +152,14 @@ describe('Dashboard Collections Page', () => {
     });
 
     it('should select a different collection', () => {
-      let name = 'http_testcollection';
-      let version = '001';
+      const name = 'http_testcollection';
+      const version = '001';
 
       // First visit the collections page in order to fetch the list of
       // collections with which to populate the dropdown on the collection
       // details page.
       cy.visit('/collections');
+      cy.clearStartDateTime();
       cy.wait('@getCollections');
       cy.get('.table .tbody .tr').should('have.length', 5);
 
@@ -193,7 +226,7 @@ describe('Dashboard Collections Page', () => {
 
       // update collection and submit
       const duplicateHandling = 'version';
-      const meta = {metaObj: 'metadata'};
+      const meta = { metaObj: 'metadata' };
       cy.contains('.ace_variable', 'name');
       cy.editJsonTextarea({ data: { duplicateHandling, meta }, update: true });
       cy.contains('form button', 'Submit').click();
@@ -243,6 +276,7 @@ describe('Dashboard Collections Page', () => {
       cy.route('DELETE', '/collections/https_testcollection/001').as('deleteCollection');
 
       cy.visit(`/collections/collection/${name}/${version}`);
+      cy.clearStartDateTime();
 
       // delete collection
       cy.get('.DeleteCollection > .button').click();
@@ -288,7 +322,7 @@ describe('Dashboard Collections Page', () => {
           // before these all show up or don't show up correctly.
           cy.get(
             `[data-value="${collection.name}___${collection.version}"] > .table__main-asset > a`,
-            {timeout: 25000}).should(existOrNotExist);
+            { timeout: 25000 }).should(existOrNotExist);
         });
       cy.get('.table .tbody .tr').should('have.length', 4);
       cy.task('resetState');
@@ -301,6 +335,7 @@ describe('Dashboard Collections Page', () => {
       cy.route('DELETE', '/collections/MOD09GK/006').as('deleteCollection');
 
       cy.visit(`/collections/collection/${name}/${version}`);
+      cy.clearStartDateTime();
 
       // delete collection
       cy.get('.DeleteCollection > .button').click();